split up subscription manager
