+++ src/page/managers/tagManager/TagManager.ts
@@ -2,9 +2,9 @@ import type {
   TagsObjectForApi,
   TagsObjectWithBoolean,
 } from 'src/page/tags/types';
+import type { ContextInterface } from 'src/shared/context/types';
 import Log from '../../../shared/libraries/Log';
 import TagUtils from '../../../shared/utils/TagUtils';
-import type { ContextInterface } from '../../models/Context';
 import type { ITagManager } from './types';
 
 /**
