+++ src/shared/helpers/init.ts
@@ -1,6 +1,6 @@
 import Bell from '../../page/bell/Bell';
-import type { ContextInterface } from '../../page/models/Context';
 import type { AppConfig } from '../config/types';
+import type { ContextInterface } from '../context/types';
 import Log from '../libraries/Log';
 import { CustomLinkManager } from '../managers/CustomLinkManager';
 import { NotificationPermission } from '../models/NotificationPermission';
