+++ src/shared/managers/subscription/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_DEVICE_ID = '99999999-9999-9999-9999-999999999999';
